refactor: share tool mutation classification
